redo lua folder structure
