+++ config/fxdata/lua/config-api/templates.lua
