+++ config/fxdata/lua/managers/MapManager.lua
@@ -1,0 +1,2 @@
+-- MapManager.lua
+-- Utility module for map manipulation (reveal/conceal tiles, spawn entities, change slabs).
